Submission rewrite in progress
Most stuff is broken.
The data from the provenance step gets saved to a json column in the
new labwares table, and read back out to be inserted back into the
form if you come back to the page.
Multiple labware doesn't quite work: possibly the javascript needs
to be a bit cleverer figuring out which fields need to be filled.
The structure of the form is generated from the materials schema
given by the MatconClient.