submissions_controller - make sure submission is ok before trying to dispatch
If we save incomplete data in the provenance step, make sure the
status is set back to 'provenance' (if it was at 'dispatch') in case
users go back and incorrect previously acceptable data.
Check that the submission has reached the 'dispatch' step before
embarking on a 'dispatch' operation. There is nothing to stop a user
trying to skip to the last step.