UI fix for claim page
When turbolinks reload the claim page, the js was recreating
both tables on the page as if they were the collections table,
so the left table (submissions) stopped working. Now that only
is done on the right table.
But the left table still has to be bootstrapified on
turbolinks load, so we added a call to bootstrapTable() for
that table, just to make it bootstrapy.