A211 - integrate with new test ehmdmc service
HMDMCs are now expected to be ##/#### (2/4 instead of 2/3). Accept
either 2/3 or 2/4 as correctly formed, for now.
There is a new test eHMDMC server. The url for this replaces
a localhost url in the development.rb config.
The new eHMDMC server returns a block of json specifying if the
HMDMC code is valid or not. The ehmdmc_client has been
correspondingly altered.